Research-backed rewrite: float32 + patched upstream + MPS Metal kernel
Based on 11 research docs, Codex + Gemini reviews:

Upstream pipeline patch (patches/001-fix-hardcoded-bfloat16.patch):
- Replace hardcoded torch.bfloat16 with self.dtype (lines 586, 600)
- Make _autocast_context use self.dtype instead of hardcoded bfloat16
- These 3 changes fix the root cause of MPS dtype assertion

Engine rewrite:
- Use float32 on MPS (research: faster than float16, 18-20s vs 22-25s)
- Load on CPU first with low_cpu_mem_usage=True
- Quantize BEFORE moving to device (saves memory during transfer)
- Remove broken sequential/model CPU offload

MPS environment:
- Add PYTORCH_MPS_PREFER_METAL=1 (stable Metal matmul kernel)
- Add PYTORCH_MPS_FAST_MATH=1 (faster Metal shaders)

All 13/13 tests pass.

+++ tests/test_optimizations.py
@@ -129,7 +129,8 @@ class TestMPSBackend:
         import torch
         from realrestore_cli.optimizations.mps_backend import get_optimal_dtype
         dtype = get_optimal_dtype()
-        assert dtype == torch.float16
+        # Research: float32 is faster than float16 on Apple Silicon
+        assert dtype == torch.float32
 
     def test_hardware_info(self):
         from realrestore_cli.optimizations.mps_backend import get_apple_silicon_info
